Debugging an error.
The error: TypeError: argument of type 'NoneType' is not iterable.
The error occures because a value was added to the Variety_ID column (in the Wine table) at the end,
and the rest of the columns take the value None.
When the loop goes through all the base records in the table, it comes across a new record with the value None.